requirements.txt
Flask[async]==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
Werkzeug==3.0.3
click==8.1.7
reportlab==4.2.0
qrcode==7.4.2
Pillow==10.3.0
opencv-python==4.9.0.80
requests==2.32.3
python-dotenv==1.0.1
numpy==1.26.4
pytesseract==0.3.12
pymupdf==1.24.0